// ui/src/helpers/config-diff/types.ts
/*
 * Copyright (C) 2026 The ORT Server Authors (See <https://github.com/eclipse-apoapsis/ort-server/blob/main/NOTICE>)
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     https://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 *
 * SPDX-License-Identifier: Apache-2.0
 * License-Filename: LICENSE
 */

export type RunConfigurationDiffStatus = 'added' | 'removed' | 'modified';

export type RunConfigurationDiffPathSegment = string | number;

type BaseRunConfigurationDiffEntry = {
  path: string;
  pathSegments: RunConfigurationDiffPathSegment[];
};

export type RunConfigurationDiffAddedEntry = BaseRunConfigurationDiffEntry & {
  status: 'added';
  newValue: unknown;
};

export type RunConfigurationDiffRemovedEntry = BaseRunConfigurationDiffEntry & {
  status: 'removed';
  oldValue: unknown;
};

export type RunConfigurationDiffModifiedEntry =
  BaseRunConfigurationDiffEntry & {
    status: 'modified';
    oldValue: unknown;
    newValue: unknown;
  };

export type RunConfigurationDiffEntry =
  | RunConfigurationDiffAddedEntry
  | RunConfigurationDiffRemovedEntry
  | RunConfigurationDiffModifiedEntry;

export type RunConfigurationDiffCounts = {
  added: number;
  removed: number;
  modified: number;
  total: number;
};

export type RunConfigurationDiff = {
  added: RunConfigurationDiffAddedEntry[];
  removed: RunConfigurationDiffRemovedEntry[];
  modified: RunConfigurationDiffModifiedEntry[];
  counts: RunConfigurationDiffCounts;
};